feat: rename package to gns3fy-next v1.0.0
- Renamed package from gns3fy to gns3fy-next
- Updated version to 1.0.0
- Renamed source directory: gns3fy/ -> gns3fy_next/
- Updated all imports and documentation
- Added GNS3 API v3 support (JWT authentication)
- Added project locking operations
- Added create_drawing method
- Enhanced nodes_inventory with more details
- Added GitHub Actions workflow for PyPI publishing
