+++ gns3fy_next/drawing_utils.py
