+++ pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
-name = "gns3fy"
-version = "0.7.2"
-description = "Python wrapper around GNS3 Server API"
+name = "gns3fy-next"
+version = "1.0.0"
+description = "Python wrapper around GNS3 Server API v2 and v3"
 authors = ["David Flores <davidflores7_8@hotmail.com>"]
 license = "MIT"
 readme = "README.md"
@@ -19,6 +19,8 @@ classifiers = [
 python = "^3.8"
 requests = "^2.22"
 pydantic = "^2.0"
+pyjwt = "^2.0"
+urllib3 = "^1.26"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^5.0"
