+++ tests/__init__.py
@@ -1,4 +1,4 @@
 import sys
 import os
 
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "../gns3fy")))
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "../gns3fy_next")))
